requirements.txt
blinker==1.3
fake-factory==0.5.3
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
geolocation-python==0.2.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
requests==2.8.1
SQLAlchemy==1.0.3
Werkzeug==0.10.4
wheel==0.24.0
